Move workflow task to datasets module;
