+++ services/src/datasets/mod.rs
@@ -1,6 +1,12 @@
 pub mod add_from_directory;
+mod create_from_workflow;
 pub mod external; // TODO: move to layers/external
 pub mod in_memory;
 pub mod listing;
 pub mod storage;
 pub mod upload;
+
+pub(crate) use create_from_workflow::{
+    schedule_raster_dataset_from_workflow_task, RasterDatasetFromWorkflow,
+    RasterDatasetFromWorkflowResult,
+};
