+++ services/src/handlers/workflows.rs
@@ -1,43 +1,23 @@
 use std::collections::HashSet;
 use std::io::{Cursor, Write};
-use std::path::PathBuf;
 
-use crate::api::model::datatypes::{DataId, DatasetId};
-use crate::api::model::services::AddDataset;
+use crate::api::model::datatypes::DataId;
 use crate::datasets::listing::{DatasetProvider, ProvenanceOutput};
-use crate::datasets::storage::{DatasetDefinition, DatasetStore, MetaDataDefinition};
-use crate::datasets::upload::{UploadId, UploadRootPath};
 use crate::error::Result;
 use crate::handlers::Context;
 use crate::layers::storage::LayerProviderDb;
-use crate::util::config::get_config_element;
-use crate::util::user_input::UserInput;
 use crate::util::IdResponse;
 use crate::workflows::registry::WorkflowRegistry;
 use crate::workflows::workflow::{Workflow, WorkflowId};
 use actix_web::{web, FromRequest, HttpResponse, Responder};
 use futures::future::join_all;
 use geoengine_datatypes::error::{BoxedResultExt, ErrorSource};
-use geoengine_datatypes::primitives::RasterQueryRectangle;
-use geoengine_datatypes::spatial_reference::SpatialReference;
-use geoengine_datatypes::util::Identifier;
-use geoengine_operators::engine::{
-    ExecutionContext, OperatorData, TypedOperator, TypedResultDescriptor,
-};
-use geoengine_operators::source::{
-    GdalLoadingInfoTemporalSlice, GdalMetaDataList, GdalMetaDataStatic,
-};
-use geoengine_operators::util::raster_stream_to_geotiff::{
-    raster_stream_to_geotiff, GdalGeoTiffDatasetMetadata, GdalGeoTiffOptions,
-};
-use geoengine_operators::{call_on_generic_raster_processor_gdal_types, call_on_typed_operator};
+use geoengine_operators::call_on_typed_operator;
+use geoengine_operators::engine::{OperatorData, TypedOperator, TypedResultDescriptor};
 
+use crate::datasets::{schedule_raster_dataset_from_workflow_task, RasterDatasetFromWorkflow};
 use crate::handlers::tasks::TaskResponse;
-use crate::tasks::{Task, TaskManager, TaskStatusInfo};
-use serde::{Deserialize, Serialize};
 use snafu::{ResultExt, Snafu};
-use tokio::fs;
-use utoipa::ToSchema;
 use zip::{write::FileOptions, ZipWriter};
 
 pub(crate) fn init_workflow_routes<C>(cfg: &mut web::ServiceConfig)
@@ -358,192 +338,6 @@ async fn resolve_provenance<C: Context>(
     }
 }
 
-/// parameter for the dataset from workflow handler (body)
-#[derive(Clone, Debug, Deserialize, Serialize, ToSchema)]
-#[schema(example = json!({"name": "foo", "description": null, "query": {"spatialBounds": {"upperLeftCoordinate": {"x": -10.0, "y": 80.0}, "lowerRightCoordinate": {"x": 50.0, "y": 20.0}}, "timeInterval": {"start": 1_388_534_400_000_i64, "end": 1_388_534_401_000_i64}, "spatialResolution": {"x": 0.1, "y": 0.1}}}))]
-pub struct RasterDatasetFromWorkflow {
-    name: String,
-    description: Option<String>,
-    query: RasterQueryRectangle,
-    #[schema(default = default_as_cog)]
-    #[serde(default = "default_as_cog")]
-    as_cog: bool,
-}
-
-/// By default, we set [`RasterDatasetFromWorkflow::as_cog`] to true to produce cloud-optmized `GeoTiff`s.
-#[inline]
-const fn default_as_cog() -> bool {
-    true
-}
-
-/// response of the dataset from workflow handler
-#[derive(Clone, Debug, Deserialize, Serialize, ToSchema)]
-pub struct RasterDatasetFromWorkflowResult {
-    dataset: DatasetId,
-    upload: UploadId,
-}
-
-impl TaskStatusInfo for RasterDatasetFromWorkflowResult {}
-
-async fn create_dataset<C: Context>(
-    info: RasterDatasetFromWorkflow,
-    mut slice_info: Vec<GdalLoadingInfoTemporalSlice>,
-    result_descriptor: &geoengine_operators::engine::RasterResultDescriptor,
-    ctx: &C,
-    session: <C as Context>::Session,
-) -> Result<DatasetId> {
-    let dataset_id = DatasetId::new();
-    //TODO: Recognize MetaDataDefinition::GdalMetaDataRegular
-    let meta_data = if slice_info.len() == 1 {
-        let loading_info_slice = slice_info.pop().expect("slice_info has len one");
-        let time = Some(loading_info_slice.time);
-        let params = loading_info_slice
-            .params
-            .expect("datasets with exactly one timestep should have data");
-        MetaDataDefinition::GdalStatic(GdalMetaDataStatic {
-            time,
-            params,
-            result_descriptor: result_descriptor.clone(),
-        })
-    } else {
-        MetaDataDefinition::GdalMetaDataList(GdalMetaDataList {
-            result_descriptor: result_descriptor.clone(),
-            params: slice_info,
-        })
-    };
-
-    let dataset_definition = DatasetDefinition {
-        properties: AddDataset {
-            id: Some(dataset_id),
-            name: info.name,
-            description: info.description.unwrap_or_default(),
-            source_operator: "GdalSource".to_owned(),
-            symbology: None,  // TODO add symbology?
-            provenance: None, // TODO add provenance that references the workflow
-        },
-        meta_data,
-    };
-
-    // TODO: build pyramides, prefereably in the background
-
-    let db = ctx.dataset_db_ref();
-    let meta = db.wrap_meta_data(dataset_definition.meta_data);
-    let dataset = db
-        .add_dataset(&session, dataset_definition.properties.validated()?, meta)
-        .await?;
-
-    Ok(dataset)
-}
-
-struct RasterDatasetFromWorkflowTask<C: Context> {
-    id: web::Path<WorkflowId>,
-    session: C::Session,
-    ctx: web::Data<C>,
-    info: web::Json<RasterDatasetFromWorkflow>,
-    upload: UploadId,
-    file_path: PathBuf,
-}
-
-impl<C: Context> RasterDatasetFromWorkflowTask<C> {
-    async fn process(&self) -> Result<RasterDatasetFromWorkflowResult> {
-        let workflow = self.ctx.workflow_registry_ref().load(&self.id).await?; //TODO: Maybe task should not start if it would fail here.
-
-        let operator = workflow
-            .operator
-            .get_raster()
-            .context(crate::error::Operator)?;
-
-        let execution_context = self.ctx.execution_context(self.session.clone())?;
-        let initialized = operator
-            .clone()
-            .initialize(&execution_context)
-            .await
-            .context(crate::error::Operator)?;
-
-        let result_descriptor = initialized.result_descriptor();
-
-        let processor = initialized
-            .query_processor()
-            .context(crate::error::Operator)?;
-
-        let query_rect = self.info.query;
-        let query_ctx = self.ctx.query_context(self.session.clone())?;
-        let request_spatial_ref =
-            Option::<SpatialReference>::from(result_descriptor.spatial_reference)
-                .ok_or(crate::error::Error::MissingSpatialReference)?;
-        let tile_limit = None; // TODO: set a reasonable limit or make configurable?
-
-        // build the geotiff
-        let res = call_on_generic_raster_processor_gdal_types!(processor, p => raster_stream_to_geotiff(
-            &self.file_path,
-            p,
-            query_rect,
-            query_ctx,
-            GdalGeoTiffDatasetMetadata {
-                no_data_value: Default::default(), // TODO: decide how to handle the no data here
-                spatial_reference: request_spatial_ref,
-            },
-            GdalGeoTiffOptions {
-                compression_num_threads: get_config_element::<crate::util::config::Gdal>()?.compression_num_threads,
-                as_cog: self.info.as_cog,
-                force_big_tiff: false,
-            },
-            tile_limit,
-            Box::pin(futures::future::pending()), // datasets shall continue to be built in the background and not cancelled
-            execution_context.tiling_specification(),
-        ).await)?
-            .map_err(crate::error::Error::from)?;
-
-        // create the dataset
-        let dataset = create_dataset(
-            self.info.clone(),
-            res,
-            result_descriptor,
-            self.ctx.get_ref(),
-            self.session.clone(),
-        )
-        .await?;
-
-        Ok(RasterDatasetFromWorkflowResult {
-            dataset,
-            upload: self.upload,
-        })
-    }
-}
-
-#[async_trait::async_trait]
-impl<C: Context> Task<C::TaskContext> for RasterDatasetFromWorkflowTask<C> {
-    async fn run(
-        &self,
-        _ctx: C::TaskContext,
-    ) -> Result<Box<dyn crate::tasks::TaskStatusInfo>, Box<dyn ErrorSource>> {
-        let response = self.process().await;
-
-        response
-            .map(TaskStatusInfo::boxed)
-            .map_err(ErrorSource::boxed)
-    }
-
-    async fn cleanup_on_error(&self, _ctx: C::TaskContext) -> Result<(), Box<dyn ErrorSource>> {
-        fs::remove_dir_all(&self.file_path)
-            .await
-            .context(crate::error::Io)
-            .map_err(ErrorSource::boxed)?;
-
-        //TODO: Dataset might already be in the database, if task was already close to finishing.
-
-        Ok(())
-    }
-
-    fn task_type(&self) -> &'static str {
-        "create-dataset"
-    }
-
-    fn task_unique_id(&self) -> Option<String> {
-        Some(self.upload.to_string())
-    }
-}
-
 /// Create a task for creating a new dataset from the result of the workflow given by its `id` and the dataset parameters in the request body.
 /// Returns the id of the created task
 #[utoipa::path(
@@ -569,26 +363,13 @@ async fn dataset_from_workflow_handler<C: Context>(
     ctx: web::Data<C>,
     info: web::Json<RasterDatasetFromWorkflow>,
 ) -> Result<impl Responder> {
-    let inner_context = ctx.clone().into_inner();
-
-    let upload = UploadId::new();
-    let upload_path = upload.root_path()?;
-    fs::create_dir_all(&upload_path)
-        .await
-        .context(crate::error::Io)?;
-    let file_path = upload_path.clone();
-
-    let task = RasterDatasetFromWorkflowTask {
-        id,
+    let task_id = schedule_raster_dataset_from_workflow_task(
+        id.into_inner(),
         session,
-        ctx: ctx.clone(),
-        info,
-        upload,
-        file_path,
-    }
-    .boxed();
-
-    let task_id = inner_context.tasks_ref().schedule(task, None).await?;
+        ctx.into_inner(),
+        info.into_inner(),
+    )
+    .await?;
 
     Ok(web::Json(TaskResponse::new(task_id)))
 }
@@ -611,9 +392,11 @@ mod tests {
 
     use super::*;
     use crate::contexts::{InMemoryContext, Session, SimpleContext};
+    use crate::datasets::RasterDatasetFromWorkflowResult;
     use crate::handlers::ErrorResponse;
     use crate::tasks::util::test::wait_for_task_to_finish;
-    use crate::tasks::{TaskCleanUpStatus, TaskStatus};
+    use crate::tasks::{TaskCleanUpStatus, TaskManager, TaskStatus};
+    use crate::util::config::get_config_element;
     use crate::util::tests::{
         add_ndvi_to_datasets, check_allowed_http_methods, check_allowed_http_methods2,
         read_body_string, register_ndvi_workflow_helper, send_test_request, TestDataUploads,
@@ -625,14 +408,14 @@ mod tests {
     use actix_web_httpauth::headers::authorization::Bearer;
     use geoengine_datatypes::collections::MultiPointCollection;
     use geoengine_datatypes::primitives::{
-        ContinuousMeasurement, FeatureData, Measurement, MultiPoint, SpatialPartition2D,
-        SpatialResolution, TimeInterval,
+        ContinuousMeasurement, FeatureData, Measurement, MultiPoint, RasterQueryRectangle,
+        SpatialPartition2D, SpatialResolution, TimeInterval,
     };
     use geoengine_datatypes::raster::{GridShape, RasterDataType, TilingSpecification};
     use geoengine_datatypes::spatial_reference::SpatialReference;
     use geoengine_datatypes::util::test::TestDefault;
     use geoengine_operators::engine::{
-        MultipleRasterOrSingleVectorSource, PlotOperator, TypedOperator,
+        ExecutionContext, MultipleRasterOrSingleVectorSource, PlotOperator, TypedOperator,
     };
     use geoengine_operators::engine::{RasterOperator, RasterResultDescriptor, VectorOperator};
     use geoengine_operators::mock::{
@@ -642,7 +425,10 @@ mod tests {
     use geoengine_operators::plot::{Statistics, StatisticsParams};
     use geoengine_operators::source::{GdalSource, GdalSourceParameters};
     use geoengine_operators::util::input::MultiRasterOrVectorOperator::Raster;
-    use geoengine_operators::util::raster_stream_to_geotiff::single_timestep_raster_stream_to_geotiff_bytes;
+    use geoengine_operators::util::raster_stream_to_geotiff::{
+        single_timestep_raster_stream_to_geotiff_bytes, GdalGeoTiffDatasetMetadata,
+        GdalGeoTiffOptions,
+    };
     use serde_json::json;
     use std::io::Read;
     use zip::read::ZipFile;
